@@ -6,4 +6,4 @@
 <h2>@ViewData["Title"]</h2>
 <h3>@ViewData["Message"]</h3>
 
-<p>@Localizer["about.infoText"]</p>
+<p>@Localizer["welcomeToReact"]</p>
